QemuOpenBoardPkg: Add QemuOpenBoardPkg (Stage 1)
QemuOpenBoardPkg adds a MinPlatform port to QEMU x86_64.
This package can boot UEFI Linux and Windows.
This port brings a starting place for understanding the MinPlatform,
and board porting.
QemuOpenBoardPkg is my project for Google Summer of Code 2022.
I proposed this project to better understand UEFI inner workings,
and to get started with platform level code.

This patch add the base for QemuOpenBoardPkg.
It also enable MinPlatform stage 1.

Signed-off-by: Théo Jehl <theojehl76@gmail.com>